(feat): cleanup

+++ next2ts.js
@@ -1,26 +1,30 @@
-//TODO: Run npm install
+// https://gist.github.com/scriptex/20536d8cda36221f91d69a6bd4a528b3
+
 //TODO: check for the project page directory
 //TODO: Rename all files in the tree
+//TODO: Run npm install
 
+import spawn from 'cross-spawn';
 import fs from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-// function folderExists(path) {
-//   return new Promise((resolve, reject) => {
-//     access(path, function (err) {
-//       if (err) {
-//         if (err.code === 'ENOENT') {
-//           return resolve(false);
-//         }
-//         return reject(err);
-//       }
-//       return resolve(true);
-//     });
-//   });
-// }
+
+function folderExists(path) {
+  return new Promise((resolve, reject) => {
+    access(path, function (err) {
+      if (err) {
+        if (err.code === 'ENOENT') {
+          return resolve(false);
+        }
+        return reject(err);
+      }
+      return resolve(true);
+    });
+  });
+}
 
 function hasYarn() {
   return fs.existsSync(path.resolve(process.cwd(), 'yarn.lock'));
@@ -30,6 +34,17 @@ export async function init() {
   const root = path.basename(__dirname);
 
   console.log(`\nMigrating ${root} to typescript...`);
+
+  if (
+    !folderExists(path.join(__dirname + '/pages')) ||
+    !folderExists(path.join(__dirname + '/src/pages'))
+  ) {
+    throw new Error('This is not a next.js project');
+  }
+
+  let projectSource = folderExists(path.join(__dirname + '/pages'))
+    ? path.join(__dirname + '/pages')
+    : path.join(__dirname + '/src/pages');
 
   try {
     fs.copyFileSync(
@@ -44,59 +59,9 @@ export async function init() {
 
   const allDependancies = ['typescript', '@types/react', '@types/node'];
 
-  // if (hasYarn()) {
-  //   console.log('Installing packages with yarn');
-  // } else {
-  //   console.log('Installing package with npm');
-  // }
-
-  // const baseName = path.basename(__dirname);
-
-  // let pagesExists = await folderExists(join(__dirname, 'pages/index.js'));
-  // let pagesExistsInSrc = await folderExists(
-  //   join(__dirname, 'src/pages/index.js')
-  // );
-
-  // if (!pagesExists && !pagesExistsInSrc) {
-  //   console.log('\n');
-  //   console.log(chalk.yellowBright('next2ts\n'));
-  //   console.log(chalk.red(`${baseName} is not a Next.js Project!`));
-  //   process.exit(0);
-  // }
-
-  // if (hasYarn()) {
-  //   //execSync('yarn add --dev typescript @types/react @types/node');
-  // } else {
-  //   //execSync('npm install -D typescript @types/react @types/node');
-  // }
-  //execSync('touch tsconfig.json');
-
-  // fs.writeFile(
-  //   'tsconfig.json',
-  //   `
-  // {
-  //   "target": "es5",
-  //   "lib": ["dom", "dom.iterable", "esnext"],
-  //   "allowJs": true,
-  //   "skipLibCheck": true,
-  //   "strict": false,
-  //   "forceConsistentCasingInFileNames": true,
-  //   "noEmit": true,
-  //   "esModuleInterop": true,
-  //   "module": "esnext",
-  //   "moduleResolution": "node",
-  //   "resolveJsonModule": true,
-  //   "isolatedModules": true
-  // }
-  //     `,
-  // function (e) {
-  //   if (e) throw e;
-  //   console.log(chalk.blue(`Added tsconfig.json in the root`));
-  // }
-  // );
-
-  //TODO: touch tsconfig.json
-  //TODO: Rename file and folders to tsx
+  return install(root, hasYarn(), allDependancies).then(() => {
+    console.log('Installing...............');
+  });
 }
 
 function install(root, useYarn, dependancies) {
@@ -115,5 +80,17 @@ function install(root, useYarn, dependancies) {
         dependancies
       );
     }
+
+    const child = spawn(command, args, { stdio: 'inherit' });
+
+    child.on('close', (code) => {
+      if (code !== 0) {
+        reject({
+          command: `${command} ${args.join(' ')}`,
+        });
+        return;
+      }
+      resolve();
+    });
   });
 }
